refactor: se realizaron cambios componenteB, en pom, se agregró plugin y dependencias de OpenApi, se cambiaron controladores, modelos, repositorios y servicios para que se pueda generar las especificaciones de apis en base al archivo YAML, tambien se hicieron cambios leves al YAML

+++ componenteB/src/main/java/uk/techiegt/profa/componenteB/entities/Factura.java
@@ -6,8 +6,8 @@ import lombok.*;
 @Entity @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
 public class Factura {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
-    private Long proveedorId;
+    private Integer proveedorId;
     private double totalFactura;
 }
