+++ componenteB/src/main/java/uk/techiegt/profa/componenteB/repositories/ProveedorRepository.java
@@ -3,4 +3,4 @@ package uk.techiegt.profa.componenteB.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uk.techiegt.profa.componenteB.entities.Proveedor;
 
-public interface ProveedorRepository extends JpaRepository<Proveedor, Long> {}
+public interface ProveedorRepository extends JpaRepository<Proveedor, Integer> {}
